% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{compute_z_score.default}
\alias{compute_z_score.default}
\title{compute_z_score}
\usage{
\method{compute_z_score}{default}(par, fit.stan)
}
\description{
Calculate z-score of parameter estimates from a vector using the
the true posterior mean and standard deviation (based on MCMC).
}

