% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{compute_z_score.gmo}
\alias{compute_z_score.gmo}
\title{compute_z_score}
\usage{
\method{compute_z_score}{gmo}(fit.gmo, fit.stan)
}
\arguments{
\item{gmo_obj}{A gmo object with estimated hyperparameters.}

\item{stanfit}{A stanfit object with MCMC samples.}
}
\value{
For i=1,...,length(phi),
  z_i = ( (phi^{GMO}_i - mean^{MCMC}(phi_i) ) / sd^{MCMC}(phi_i)
}
\description{
Calculate z-score of parameter estimates from GMO using the
the true posterior mean and standard deviation (based on MCMC).
}

