% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_covariance.R
\name{est_covariance}
\alias{est_covariance}
\title{Uncertainty estimate of marginal posterior parameters}
\usage{
est_covariance(fit.gmo, ...)
}
\arguments{
\item{fit.gmo}{An object of reference class \code{"gmo"}.}
}
\description{
Calculate a covariance matrix of the marginal posterior
parameters, based on an unbiased estimate of the inverse of the
observed Fisher information.
}

