% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{class}
\name{opt-class}
\alias{Opt}
\alias{opt-class}
\title{A Reference Class for optimization.}
\description{
A Reference Class for optimization.
}
\section{Fields}{

\describe{
\item{\code{eta}}{Double, constant scale factor for learning rate.}

\item{\code{grad_prev}}{A numeric vector, previous gradient value.}

\item{\code{I_est}}{A numeric vector, an estimate of the inverse Fisher information.}
}}
\section{Methods}{

\describe{
\item{\code{update_params(par, grad, iter)}}{Update parameters using Nesterov's accelerated gradient with
RMSprop.}
}}

