% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gmo.R
\name{print.gmo}
\alias{print.gmo}
\title{Print objects of class \code{gmo}.}
\usage{
\method{print}{gmo}(x, ...)
}
\arguments{
\item{x}{object of class \code{gmo}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Print objects of class \code{gmo}.
}

